import use-cases from github
the rationale of this PR is to seed the Use Case documents with what we have on github.

It is expected that the editors then
- move each use-case into the appropriate section (Functional, Non-Functional, ...)
- replace the `<div class="issue">` with the appopriate text (possibly reusing text from the github issue)

The id of each issue (uc{issue_number}) is expected to remain valid,
so that people can start pointing at use-cases in the document.
If two use-cases are merged as duplicate,
anchors (`<a name="uc{issue_number}">`) should be added in the appropriate section,
to avoid breaking links.

----

For the record, the markdown file `spec\pending_usecases.md`
was generated with the following command:
```sh
gh issue list --label usecase -L 1000 --search "sort:created-asc" --json title,number --template '
<section data-max-toc="2">
<h2>Pending use cases</h2>

The following use-cases are pending triage into one of the previous sections.

{{range .}}
<section id="uc{{.number}}">
<h3>{{.title}}</h3>
<div class=issue data-number={{.number}}></div>
</section>

{{end}}
</section>
' | sed 's/<h3>\[UC\] /<h3>/' >spec/pending_usecases.md
```